refactor: one file tree component for the dock and the pull request
PullsFileTree was FilesPanel's TreeRow with a selection highlight bolted on.
Both now render FileTree, which takes the highlight and a defaultOpen flag —
the pull request's changed files start expanded, the repo browser collapsed —
and tracks folders toggled away from that default, so a folder's state
survives its parent closing.
